move handler to apiserver
